Karukaeru

+++ ryohin/src/main/java/com/example/ryohin/dto/product/ProductDetail.java
@@ -4,7 +4,7 @@ package com.example.ryohin.dto.product;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.math.BigDecimal;
+
 
 
 @Data
@@ -15,7 +15,7 @@ public class ProductDetail {
     private String productName;
     private String description;
     private String material;
-    private BigDecimal price;
+    private int price;
     private String imageUrl;
     private Integer stockQuantity;
 }
