+++ ryohin/src/main/java/com/example/ryohin/dto/product/ProductListItem.java
@@ -3,7 +3,7 @@ package com.example.ryohin.dto.product;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.math.BigDecimal;
+
 
 @Data
 @NoArgsConstructor
@@ -12,5 +12,5 @@ public class ProductListItem {
     private Integer productId;
     private String productName;
     private String imageUrl;
-    private BigDecimal price;
+    private int price;
 }
